# prop/portal/openapi/prop-api.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Prop.Api | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/portal/admin/accounts": {
      "post": {
        "tags": [
          "Portal"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartAccountRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountResponse"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Portal"
        ],
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "group",
            "in": "query",
            "schema": {
              "default": "All",
              "$ref": "#/components/schemas/AccountGroup"
            }
          },
          {
            "name": "challengeId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "before",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminAccountsResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/portal/admin/accounts/{accountId}": {
      "get": {
        "tags": [
          "Portal"
        ],
        "parameters": [
          {
            "name": "accountId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountDetailsResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AccountCountsResponse": {
        "required": [
          "all",
          "evaluation",
          "awaitingFunding",
          "funded",
          "ended"
        ],
        "type": "object",
        "properties": {
          "all": {
            "type": "integer",
            "format": "int32"
          },
          "evaluation": {
            "type": "integer",
            "format": "int32"
          },
          "awaitingFunding": {
            "type": "integer",
            "format": "int32"
          },
          "funded": {
            "type": "integer",
            "format": "int32"
          },
          "ended": {
            "type": "integer",
            "format": "int32"
          }
        },
        "description": "How many accounts are in each group."
      },
      "AccountDetailsResponse": {
        "required": [
          "account",
          "challenge",
          "live",
          "stages",
          "results",
          "breach",
          "expiry",
          "endedAt",
          "payouts",
          "historyVersion"
        ],
        "type": "object",
        "properties": {
          "account": {
            "$ref": "#/components/schemas/AccountResponse"
          },
          "challenge": {
            "$ref": "#/components/schemas/ChallengeDefinition"
          },
          "live": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/LiveFigures"
              }
            ]
          },
          "stages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StageResponse"
            }
          },
          "results": {
            "$ref": "#/components/schemas/ResultsResponse"
          },
          "breach": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/BreachEvidence"
              }
            ]
          },
          "expiry": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ExpiryEvidence"
              }
            ]
          },
          "endedAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "payouts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PayoutResponse"
            }
          },
          "historyVersion": {
            "type": "string"
          },
          "retry": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/RetryOffer"
              }
            ]
          }
        },
        "description": "An account as the portal shows it: the account, its trading account valued right now, the challenge it was bought\nwith, its stages, its results, the evidence if a floor was breached, why it expired if it ran out of time, when it\nended, and its payouts, newest first. HistoryVersion changes whenever the account's trading\nhistory or the rule engine's steps do, so the portal asks for the history again only then. Retry\nis how a failed challenge can be tried again."
      },
      "AccountGroup": {
        "enum": [
          "All",
          "Evaluation",
          "AwaitingFunding",
          "Funded",
          "Ended"
        ],
        "description": "Which accounts the admin panel lists. Evaluation and funded accounts are those trading or opening their trading account."
      },
      "AccountResponse": {
        "required": [
          "id",
          "number",
          "email",
          "challengeId",
          "reference",
          "status",
          "stage",
          "stageName",
          "funded",
          "tradingAccountId",
          "tradingDays",
          "minTradingDays",
          "initialBalance",
          "currency",
          "profitTarget",
          "balance",
          "openPositions",
          "dailyFloor",
          "maxLossFloor",
          "createdAt",
          "nextPayout",
          "paused",
          "stageDeadline",
          "inactivityDeadline"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "number": {
            "type": "integer",
            "format": "int64"
          },
          "email": {
            "type": "string"
          },
          "challengeId": {
            "type": "string"
          },
          "reference": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "$ref": "#/components/schemas/ChallengeStatus"
          },
          "stage": {
            "type": "integer",
            "format": "int32"
          },
          "stageName": {
            "type": "string"
          },
          "funded": {
            "type": "boolean"
          },
          "tradingAccountId": {
            "type": [
              "null",
              "string"
            ]
          },
          "tradingDays": {
            "type": "integer",
            "format": "int32"
          },
          "minTradingDays": {
            "type": "integer",
            "format": "int32"
          },
          "initialBalance": {
            "type": "number",
            "format": "double"
          },
          "currency": {
            "type": "string"
          },
          "profitTarget": {
            "type": [
              "null",
              "number"
            ],
            "format": "double"
          },
          "balance": {
            "type": [
              "null",
              "number"
            ],
            "format": "double"
          },
          "openPositions": {
            "type": "integer",
            "format": "int32"
          },
          "dailyFloor": {
            "type": [
              "null",
              "number"
            ],
            "format": "double"
          },
          "maxLossFloor": {
            "type": [
              "null",
              "number"
            ],
            "format": "double"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "nextPayout": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/PayoutQuoteResponse"
              }
            ]
          },
          "paused": {
            "type": "boolean"
          },
          "stageDeadline": {
            "type": [
              "null",
              "string"
            ],
            "format": "date"
          },
          "inactivityDeadline": {
            "type": [
              "null",
              "string"
            ],
            "format": "date"
          }
        },
        "description": "A trader's challenge account. TradingAccountId is the account on the trading platform\nfor the current stage, which starts at InitialBalance. The figures are what the trading\nplatform last reported. Paused is set while the firm's month is unpaid: no new positions\ncan be opened and the days do not count. StageDeadline is the trading day the stage fails on\nunless it is passed before, and InactivityDeadline the trading day the challenge ends on\nunless a position is opened before."
      },
...
      "AdminAccountsResponse": {
        "required": [
          "accounts",
          "counts",
          "next"
        ],
        "type": "object",
        "properties": {
          "accounts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AccountResponse"
            }
          },
          "counts": {
            "$ref": "#/components/schemas/AccountCountsResponse"
          },
          "next": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int64"
          }
        },
        "description": "A page of accounts, the counts in each group, and Next to ask for the next page with, if there is one."
      },
...
      "ChallengeStatus": {
        "enum": [
          "OpeningAccount",
          "Active",
          "AwaitingFunding",
          "Failed",
          "Cancelled"
        ]
      },
...
      "ExpiryEvidence": {
        "required": [
          "time",
          "reason",
          "day"
        ],
        "type": "object",
        "properties": {
          "time": {
            "type": "string",
            "format": "date-time"
          },
          "reason": {
            "$ref": "#/components/schemas/ExpiryReason"
          },
          "day": {
            "type": "string",
            "format": "date"
          }
        },
        "description": "The challenge ran out of time when trading day Day started, after its time limit or the days allowed without a new position."
      },
      "ExpiryReason": {
        "enum": [
          "TimeLimit",
          "Inactivity"
        ]
      },
...
      "LiveFigures": {
        "required": [
          "balance",
          "equity",
          "floors"
        ],
        "type": "object",
        "properties": {
          "balance": {
            "type": "number",
            "format": "double"
          },
          "equity": {
            "type": "number",
            "format": "double"
          },
          "floors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FloorFigure"
            }
          }
        },
        "description": "The trading account valued at the latest prices. Missing when the trading platform cannot be reached."
      },
...
      "PayoutQuoteResponse": {
        "required": [
          "canRequest",
          "refusal",
          "profit",
          "profitSplitPercent",
          "amount",
          "tradingDays",
          "minTradingDays"
        ],
        "type": "object",
        "properties": {
          "canRequest": {
            "type": "boolean"
          },
          "refusal": {
            "type": [
              "null",
              "string"
            ]
          },
          "profit": {
            "type": "number",
            "format": "double"
          },
          "profitSplitPercent": {
            "type": "number",
            "format": "double"
          },
          "amount": {
            "type": "number",
            "format": "double"
          },
          "tradingDays": {
            "type": "integer",
            "format": "int32"
          },
          "minTradingDays": {
            "type": "integer",
            "format": "int32"
          },
          "bestDayProfit": {
            "type": [
              "null",
              "number"
            ],
            "format": "double"
          },
          "consistencyPercent": {
            "type": [
              "null",
              "number"
            ],
            "format": "double"
          }
        },
        "description": "What a payout asked for now would pay the trader: ProfitSplitPercent of the profit, which\nis all withdrawn from the trading account. When CanRequest is false,\nRefusal says why. With a consistency rule, BestDayProfit is what the best\ntrading day since the last payout made, and ConsistencyPercent the most of the profit it may be.\nOnly for funded accounts."
      },
...
      "StartAccountRequest": {
        "required": [
          "email",
          "challengeId"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": [
              "null",
              "string"
            ]
          },
          "challengeId": {
            "type": [
              "null",
              "string"
            ]
          },
          "reference": {
            "type": [
              "null",
              "string"
            ]
          }
        },
        "description": "Starts a challenge for the trader with the email. Reference is the firm's own id, for\nexample its order number: repeating it returns the account already started instead of a second one."
      },
...
    }
  },
...
}